refactor: 대시보드 섹션 콘텐츠 공통으로 이동
