+++ src/features/dashboard/components/section-content/note-section-content.tsx
