+++ src/features/dashboard/components/section-content/studyroom-image-button.tsx
