+++ src/features/dashboard/components/section-content/studyroom-section-content.tsx
